Partial modulization and sync with existing setup, WIP
