+++ lua/plugins/guess-indent.lua
@@ -1,0 +1,3 @@
+return {
+  'NMAC427/guess-indent.nvim', -- Detect tabstop and shiftwidth automatically
+}
