+++ lua/plugins/init.lua
@@ -1,0 +1,3 @@
+return {
+  -- Empty, plugins managed by separate files
+}
